requirements.txt
pandas
psycopg[binary]
SQLAlchemy
python-dotenv
pytest
loguru
requests